pyproject.toml
[project]
name = "geoshader"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "db-dtypes>=1.5.1",
    "fredapi>=0.5.2",
    "google-cloud-bigquery>=3.41.0",
    "llvmlite>=0.46.0",
    "matplotlib>=3.10.8",
    "numpy>=2.4.4",
    "pandas>=2.3.3",
    "plotly>=6.6.0",
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "python-dateutil>=2.9.0.post0",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "scikit-learn>=1.8.0",
    "scipy>=1.17.1",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.3.0",
    "shap>=0.49.1",
    "statsmodels>=0.14.6",
    "torch>=2.11.0",
    "tqdm>=4.67.3",
    "yfinance>=1.2.0",
]
